Also add currency converter

+++ challenge1/src/main/java/Pants.java
@@ -1,5 +1,7 @@
+import java.math.BigDecimal;
+
 public class Pants extends Product {
-    public Pants(String productId, String name, String description, int price) {
+    public Pants(String productId, String name, String description, BigDecimal price) {
         super(productId, name, description, price);
     }
